7168a8cb489ed86b178c2f839f0681cf9eddbcf2 chore: bump version to 0.4 (志宇)
0957bb79e667edb072fc69d43a24ec1eac7a404d feat: Add `server.features` method types (志宇)
fbd8e79399df3e641ac8cef4916a2140ca63d07b feat!: Use u32 to avoid overflow (志宇)
22755f14c6ed5e31cc5db88325f2169cd7ba1d82 chore!: Bump MSRV to 1.70 (志宇)

Pull request description:

Top commit has no ACKs.

Tree-SHA512: 879c29929936046680bb963f611be6881155555da6106bf29ef5679df6b11fdf925c97ccc22e9060b7438092c4f3a21b14f487a9c7cfcf8618d750221ee1b6ce

+++ src/state.rs
@@ -60,7 +60,7 @@ impl Event {
 /// previously tracked request.
 #[derive(Debug)]
 pub struct State<PReq: PendingRequest> {
-    pending: HashMap<usize, PReq>,
+    pending: HashMap<u32, PReq>,
 }
 
 impl<PReq: PendingRequest> Default for State<PReq> {
@@ -102,13 +102,13 @@ impl<PReq: PendingRequest> State<PReq> {
     ///
     /// Returns a [`MaybeBatch<RawRequest>`], preserving whether the input was a single request or a
     /// batch.
-    pub fn track_request<R>(&mut self, next_id: &mut usize, req: R) -> MaybeBatch<RawRequest>
+    pub fn track_request<R>(&mut self, next_id: &mut u32, req: R) -> MaybeBatch<RawRequest>
     where
         R: Into<MaybeBatch<PReq>>,
     {
         fn _add_request<PReq: PendingRequest>(
             state: &mut State<PReq>,
-            next_id: &mut usize,
+            next_id: &mut u32,
             req: PReq,
         ) -> RawRequest {
             let id = *next_id;
@@ -172,14 +172,14 @@ impl<PReq: PendingRequest> State<PReq> {
 #[derive(Debug)]
 pub enum ProcessError {
     /// A response was received for an unknown or untracked request ID.
-    MissingRequest(usize),
+    MissingRequest(u32),
 
     /// The server returned a successful response, but it could not be deserialized into the
     /// expected type.
     ///
     /// The `usize` is the request ID, and the `serde_json::Error` is the underlying deserialization
     /// failure.
-    CannotDeserializeResponse(usize, serde_json::Error),
+    CannotDeserializeResponse(u32, serde_json::Error),
 
     /// A server notification could not be deserialized into the expected notification type.
     ///
